- splits the explored feed test into multiple test files based on their functionality - adds color assertion for explore feed test
