+++ app/src/androidTest/java/org/wikipedia/Constants.kt
@@ -1,0 +1,13 @@
+package org.wikipedia
+
+object Constants {
+    const val FEATURED_ARTICLE = "Featured article"
+    const val TODAY_ON_WIKIPEDIA_MAIN_PAGE = "Today on Wikipedia"
+    const val TOP_READ_ARTICLES = "Top read"
+    const val PICTURE_OF_DAY = "Picture of the day"
+    const val BECAUSE_YOU_READ = "Because you read"
+    const val NEWS_CARD = "In the news"
+    const val ON_THIS_DAY_CARD = "On this day"
+    const val RANDOM_CARD = "Random article"
+    const val SUGGESTED_EDITS = "Suggested edits"
+}
