+++ app/src/androidTest/java/org/wikipedia/TestSuite.kt
@@ -3,7 +3,6 @@ package org.wikipedia
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
 import org.junit.runners.Suite.SuiteClasses
-import org.wikipedia.tests.ExploreFeedTest
 import org.wikipedia.tests.OnboardingTest
 import org.wikipedia.tests.SuggestedEditScreenTest
 
@@ -11,6 +10,6 @@ import org.wikipedia.tests.SuggestedEditScreenTest
 @SuiteClasses(
     OnboardingTest::class,
     SuggestedEditScreenTest::class,
-    ExploreFeedTest::class,
+    ExploreFeedTestSuite::class,
 )
 class TestSuite
